[[ ssb ]] rename `platforms/ssb` to `platforms/nodejs-ssb`
refactoring platform from `ssb` to `nodejs-ssb` to prepare the ground for supporting `go-ssb` and `browsers-ssb` as well.

Related to: #104 #105 #106
