+++ src/core/platforms/nodejs-ssb/friendship.js
