+++ src/core/platforms/nodejs-ssb/rooms2.js
